move split_dld_time_from_sector_id to general loader utils, and disable linting errors
